TV Show Data Generator
This script searches for the string that the user inputs and downloads a HEADER.md file with the title, year, rating, number of seasons, overview and the trailer link. It then searches for the backdrop and creates an external folder and downloads it there, and lastly it downloads the poster and adds an entry of it to a text file from which the index page of the shows page get its list.